+++ packages/codemirror-plantuml/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @uml-drawer/codemirror-plantuml
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @uml-drawer/core@0.4.0
+
 ## 1.0.0
 
 ### Patch Changes